update for 5.0.0 Beta 7  4.3.0 Beta 8
26 February 2016, Nintendo's upcoming Sun & Moon, and announced the launch of two official Chinese version, which is the official the rectification of names Pokémon Nintendo Traditional Chinese names. In fact Nintendo has been no emphasis on the Greater China market in the past, so the name of the game there are many versions in different regions of Greater China.

In Taiwan, we used to call Pokémon to "神奇寶貝", its literal translation is "magic Baby". Yes, I also can not understand why the Taiwan publishers to take this name,but we're used to this name, so most of the people of Taiwan can not accept the new name so far.

So I have a long inner struggle, I do not know in the end or not to change to the new name Pixelmon, and now I think it should just be a transition period, so I decided to change it with Nintendo the rectification of names.

In addition to changing the name of the game from the "神奇寶貝" to "寶可夢" (transliteration of "Pokémon"), and some Pokémon name also changed (primarily in the first generation), plus I had eight months without updating it, so this update is very important in zh-TW text.